requirements/base.txt
Django==3.2.11
django-braces==1.15.0
django-extra-views==0.14.0
mysqlclient==2.1.0
python-decouple==3.5
